Move petab v2 simulation code to amici.sim
Related to #3041.

And fix type annotations.
